+++ packages/@rescript/linux-arm64/bin.d.ts
@@ -1,0 +1,13 @@
+export const binDir: string;
+
+export const binPaths: BinaryPaths;
+
+export type BinaryPaths = {
+  bsb_helper_exe: string,
+  bsc_exe: string,
+  ninja_exe: string,
+  rescript_exe: string,
+  rescript_tools_exe: string,
+  rescript_editor_analysis_exe: string,
+  rewatch_exe: string,
+};